[refactor]: OpenAI provider - move rerank types to types.go and handler to openai.go per code conventions
Rerank wire types are now exported in types.go alongside sibling
operation types, converters stay in rerank.go (embedding.go/speech.go
pattern), and HandleOpenAIRerankRequest lives in openai.go next to the
Rerank method. Documents custom-provider rerank support in
docs/providers/custom-providers.mdx and
docs/quickstart/gateway/reranking.mdx.
