+++ docs/providers/custom-providers.mdx
@@ -244,6 +244,7 @@ Available operations:
 - **`speech_stream`**: Streaming text-to-speech
 - **`transcription`**: Speech-to-text conversion
 - **`transcription_stream`**: Streaming speech-to-text
+- **`rerank`**: Document reranking via `/v1/rerank` (OpenAI-compatible base only)
 
 ### Base Provider Types
